Feature: Add -skip-non-semver flag
**cciu**'s job s to find possible update for given container images based
upon SemVer semantics. Some container images just do not use SemVer. So,
it is pointless to even try.

This commit introduces the flag `-skip-non-semver` - for tags not
following SemVer, **cciu** won't even try to check if an update is
available. Its default value is `false` to provide backward
compatibility.

Also part of this commit: various minor code-refactoring to keep code
and data apart (format strings in this case).